feat: add frame quality evaluator
